feat: update

+++ service/user/data/request/doc/doc_group.go
@@ -10,7 +10,7 @@ type CreateDocGroupRequest struct {
 
 // UpdateDocGroupRequest 更新文档分组
 type UpdateDocGroupRequest struct {
-	Id    int    `json:"id" binding:"required"`
+	Id    int64  `json:"id" binding:"required"`
 	Title string `json:"title" binding:"max=250"` // 标题
 	PId   int64  `json:"pId" binding:""`          // 父级
 	Icon  string `json:"icon" binding:""`         // 图标
